round_1/requirements.txt
Flask
Pillow
gunicorn
flask-talisman
flask-limiter